backend/requirements.txt
Flask==2.3.3
boto3>=1.34.0
python-dotenv>=1.0.0